backend/src/main/java/team/teamby/teambyteam/icalendar/application/IcalendarEventListener.java
package team.teamby.teambyteam.icalendar.application;

import lombok.RequiredArgsConstructor;
import org.springframework.scheduling.annotation.Async;
import org.springframework.stereotype.Component;
import org.springframework.transaction.event.TransactionPhase;
import org.springframework.transaction.event.TransactionalEventListener;
import team.teamby.teambyteam.icalendar.application.event.CreateIcalendarEvent;
import team.teamby.teambyteam.icalendar.util.DeployWaitingQueue;
import team.teamby.teambyteam.schedule.application.event.ScheduleEvent;
import team.teamby.teambyteam.teamplace.application.event.TeamPlaceCreatedEvent;

@Component
@RequiredArgsConstructor
public class IcalendarEventListener {

    private final DeployWaitingQueue queue;

    @Async("icalendarEventListenerAsyncExecutor")
    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
    public void createIcalendar(final TeamPlaceCreatedEvent teamPlaceCreatedEvent) {
        queue.addCreate(teamPlaceCreatedEvent.teamPlaceId());
    }

    @Async("icalendarEventListenerAsyncExecutor")
    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
    public void createIcalendar(final CreateIcalendarEvent createIcalendarEvent) {
        queue.addCreate(createIcalendarEvent.teamPlaceId());
    }

    @Async("icalendarEventListenerAsyncExecutor")
    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
    public void updateIcalendar(final ScheduleEvent scheduleEvent) {
        queue.addUpdate(scheduleEvent.getTeamPlaceId());
    }
}
